requirements.txt
CherryPy==10.1.0
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
chardet==2.3.0
cheroot==5.1.0
colorama==0.3.2
gpiozero==1.3.1
html5lib==0.999
numpy==1.8.2
pgzero==1.1
picamera==1.12
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.30
portend==1.8
pygame==1.9.2a0
pynetworktables==2017.0.7a1
pyserial==2.6
python-apt==0.9.3.12
python-debian==0.1.27
pytz==2016.10
requests==2.4.3
sense-hat==2.2.0
six==1.8.0
spidev==3.0
tempora==1.6.1
urllib3==1.9.1
wheel==0.24.0
